+++ src/components/payment/PaymentVerifyInfo.jsx
